+++ History.md
@@ -1,6 +1,8 @@
 ## [2.7.5](https://github.com/cucumber/gherkin/compare/v2.7.4...v2.7.5)
 
 * [Java] Added Formatter.close() so we can close underlying streams after a call to Formatter.done() (Aslak Hellesøy)
+* [Java] Switched from JSON simple to GSON  (Aslak Hellesøy)
+* [Java] Added a JSONPrettyFormatter  (Aslak Hellesøy)
 
 ## [2.7.4](https://github.com/cucumber/gherkin/compare/v2.7.3...v2.7.4)